Games: 8 games with Game Hub, unified leaderboard API
- Game Hub: 6 taps on avatar → game selection screen (2x4 grid)
- New games: Tetris, Minesweeper, 2048, Breakout, Flappy Bird, Helicopter
- Updated API: POST /game/score with game field, GET /game/leaderboard?game=all
- Fix: breakout int/double type, dashboard _formatDate scope
- Sync from smit_billing_app

+++ lib/screens/games/breakout_game.dart
@@ -334,8 +334,8 @@ class _BreakoutPainter extends CustomPainter {
   static const double brickOffX = _BreakoutGameState.brickOffX;
   static const double brickOffY = _BreakoutGameState.brickOffY;
   static const double brickGap = _BreakoutGameState.brickGap;
-  static const double brickCols = _BreakoutGameState.brickCols;
-  static const double brickRows = _BreakoutGameState.brickRows;
+  static const int brickCols = _BreakoutGameState.brickCols;
+  static const int brickRows = _BreakoutGameState.brickRows;
   static const double paddleW = _BreakoutGameState.paddleW;
   static const double paddleH = _BreakoutGameState.paddleH;
   static const double ballR = _BreakoutGameState.ballR;
